fairplay/spotter.py
import cv2
import hashlib
# Using standard hashing instead of raw videohash to ensure zero installation errors on diverse OS setups during a Hackathon
import os

class Spotter:
    """
    The Spotter (Asset Tracking):
    A Python-based backend continuously monitoring platforms for proprietary sports footage 
    using proxy video fingerprinting algorithms (OpenCV).
    """
    
    @staticmethod
    def generate_fingerprint(video_path):
        """Extract a stable vector hash from keyframes to detect spatial tracking matches."""
        try:
            cap = cv2.VideoCapture(video_path)
            success, frame = cap.read()
            if not success: return None
            
            # Simple perceptual hash proxy for hackathon demo: 
            # Resize frame to 8x8, convert to gray, compute mean difference
            gray = cv2.cvtColor(cv2.resize(frame, (8, 8)), cv2.COLOR_BGR2GRAY)
            avg = gray.mean()
            phash = "".join(["1" if px > avg else "0" for px in gray.flatten()])
            return hashlib.md5(phash.encode()).hexdigest()
            
        except Exception as e:
            print(f"OpenCV Tracking Error: {e}")
            return "mock_fingerprint_001A"
            
    @staticmethod
    def extract_visual_context(video_path):
        """Simulates OCR/Object detection on the video frames looking for overlays/graphics."""
        # For Hackathon context, we simulate extracting graphical overlays like tactic-lines or reaction boxes
        return "Detected: Subtitles overlay bottom-center, picture-in-picture top-right reaction cam."
