fairplay/requirements.txt
streamlit
opencv-python
videohash
ollama
supabase
python-dotenv
pandas
plotly